+++ lib/astro_chart.rb
@@ -22,6 +22,9 @@ require_relative "astro_chart/solar_return"
 require_relative "astro_chart/draconic"
 require_relative "astro_chart/dignities"
 require_relative "astro_chart/profection"
+require_relative "astro_chart/transit_timing"
+require_relative "astro_chart/solar_arc"
+require_relative "astro_chart/lunar_return"
 
 module AstroChart
 end